requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
defusedxml==0.7.1
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
httpie==3.2.1
idna==3.4
Markdown==3.4.3
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
Pygments==2.15.1
PySocks==1.7.1
pytz==2023.3
requests==2.30.0
requests-toolbelt==1.0.0
rich==13.3.5
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
